First version of inspector panel.

+++ MuseEngine/Muse/src/Core/Instrumentor.h
@@ -117,7 +117,7 @@ namespace Muse
 			long long start = std::chrono::time_point_cast<std::chrono::microseconds>(m_StartTimepoint).time_since_epoch().count();
 			long long end = std::chrono::time_point_cast<std::chrono::microseconds>(endTimepoint).time_since_epoch().count();
 
-            const uint32_t threadId = std::hash<std::thread::id>{}(std::this_thread::get_id());
+            const uint32_t threadId = static_cast<uint32_t>(std::hash<std::thread::id>{}(std::this_thread::get_id()));
 			Instrumentor::Get().WriteProfile({ m_Name, start, end, threadId });
 
 			m_Stopped = true;
